refactor(vault): split 2.4k-line store.go by concern
store.go was 2,374 lines. Relocate cohesive method groups into sibling
files within the same package (pure code movement, no logic change):

- store_records.go - record CRUD + row scan/decrypt helpers (12 funcs)
- store_grants.go  - grants + access authorization (10 funcs)
- store_crypto.go  - DEK wrap/unwrap, metadata encrypt/decrypt, password
  validation, key material (9 funcs)
- store_export.go  - vault export/import (3 funcs)

store.go keeps the Store type, constructor, vault file/catalog
management, vault CRUD, lock/unlock, and shared types (2374 -> 1001
lines). Top-level function counts unchanged.

Verified: go build ./..., go vet, gosec, and the full ./internal/vault/
test suite pass; gofmt clean.
