+++ internal/vault/store.go
@@ -3,8 +3,6 @@ package vault
 import (
 	"context"
 	"database/sql"
-	"encoding/base64"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"os"
@@ -907,1010 +905,6 @@ func (s *Store) Lock(ctx context.Context, vaultID string) error {
 	return nil
 }
 
-func (s *Store) ListRecords(ctx context.Context, filter RecordFilter, access AccessContext) ([]RecordListItem, error) {
-	if s.db == nil {
-		return nil, ErrSecretStoreUnavailable
-	}
-
-	access = access.normalized()
-	resolvedVaultID, err := s.resolveVaultID(ctx, filter.VaultID)
-	if err != nil {
-		return nil, err
-	}
-	filter.VaultID = resolvedVaultID
-	filter.WorkspaceID = strings.TrimSpace(filter.WorkspaceID)
-	filter.Type = normalizeRecordType(filter.Type)
-	if filter.Type == "*" {
-		filter.Type = ""
-	}
-	if _, err := s.getVault(ctx, filter.VaultID); err != nil {
-		return nil, err
-	}
-	vaultDB, err := s.openVaultContentDB(ctx, filter.VaultID)
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = vaultDB.Close() }()
-
-	if err := s.authorizeList(ctx, access, filter); err != nil {
-		return nil, err
-	}
-
-	items := make([]RecordListItem, 0)
-	hasRecords, err := s.hasMatchingRecords(ctx, filter)
-	if err != nil {
-		return nil, err
-	}
-	if !hasRecords {
-		s.writeAuditBestEffort(ctx, AuditEvent{
-			VaultID:     filter.VaultID,
-			WorkspaceID: access.WorkspaceID,
-			ActorType:   access.ActorType,
-			ActorID:     access.ActorID,
-			Action:      "list",
-			Outcome:     "allowed",
-			Details:     `{"count":0}`,
-		})
-		return items, nil
-	}
-
-	dek, err := s.ensureDataEncryptionKey(ctx, filter.VaultID, false)
-	if err != nil {
-		return nil, err
-	}
-
-	query := `
-		SELECT id, vault_id, type, workspace_id, source, retention_policy,
-		       metadata_nonce, metadata_ciphertext, payload_nonce, payload_ciphertext,
-		       created_at, updated_at
-		FROM vault_records
-		WHERE vault_id = ?
-		  AND (? = '' OR workspace_id = ?)
-		  AND (? = '' OR type = ?)
-		ORDER BY updated_at DESC
-	`
-	rows, err := vaultDB.QueryContext(ctx, query, filter.VaultID, filter.WorkspaceID, filter.WorkspaceID, filter.Type, filter.Type)
-	if err != nil {
-		return nil, fmt.Errorf("query vault records: %w", err)
-	}
-	defer func() { _ = rows.Close() }()
-
-	for rows.Next() {
-		row, err := scanRecordRow(rows)
-		if err != nil {
-			return nil, err
-		}
-		readCapability, _ := capabilitiesForRecordType(row.Type)
-		if access.requiresGrant() && !hasGrantWithExecutor(ctx, vaultDB, row.VaultID, access, row.WorkspaceID, readCapability, row.Type) {
-			continue
-		}
-
-		metadata, err := decryptRecordMetadata(dek, row.MetadataNonce, row.MetadataCiphertext)
-		if err != nil {
-			return nil, err
-		}
-
-		items = append(items, RecordListItem{
-			ID:              row.ID,
-			VaultID:         row.VaultID,
-			Type:            row.Type,
-			WorkspaceID:     row.WorkspaceID,
-			FolderPath:      metadata.FolderPath,
-			Label:           metadata.Label,
-			Tags:            metadata.Tags,
-			Source:          row.Source,
-			RetentionPolicy: row.RetentionPolicy,
-			CreatedAt:       row.CreatedAt,
-			UpdatedAt:       row.UpdatedAt,
-		})
-	}
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("iterate vault records: %w", err)
-	}
-
-	s.writeAuditBestEffort(ctx, AuditEvent{
-		VaultID:     filter.VaultID,
-		WorkspaceID: access.WorkspaceID,
-		ActorType:   access.ActorType,
-		ActorID:     access.ActorID,
-		Action:      "list",
-		Outcome:     "allowed",
-		Details:     fmt.Sprintf(`{"count":%d}`, len(items)),
-	})
-
-	return items, nil
-}
-
-func (s *Store) GetRecord(ctx context.Context, id string, access AccessContext) (*Record, error) {
-	return s.getRecord(ctx, id, access, false)
-}
-
-func (s *Store) getRecord(ctx context.Context, id string, access AccessContext, includeAttachmentContent bool) (*Record, error) {
-	if s.db == nil {
-		return nil, ErrSecretStoreUnavailable
-	}
-
-	id = strings.TrimSpace(id)
-	if id == "" {
-		return nil, ErrRecordNotFound
-	}
-
-	row, err := s.getRecordRow(ctx, id)
-	if err != nil {
-		return nil, err
-	}
-
-	access = access.normalized()
-	readCapability, _ := capabilitiesForRecordType(row.Type)
-	if err := s.authorizeAccess(ctx, access, row.VaultID, row.WorkspaceID, readCapability, row.Type, "read", id); err != nil {
-		return nil, err
-	}
-
-	dek, err := s.ensureDataEncryptionKey(ctx, row.VaultID, false)
-	if err != nil {
-		return nil, err
-	}
-	vaultDB, err := s.openVaultContentDB(ctx, row.VaultID)
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = vaultDB.Close() }()
-
-	record, err := decryptRecordRow(dek, row)
-	if err != nil {
-		return nil, err
-	}
-
-	attachments, err := listRecordAttachmentsWithExecutor(ctx, vaultDB, dek, row.ID, row.VaultID, includeAttachmentContent)
-	if err != nil {
-		return nil, err
-	}
-	record.Payload = mergeRecordPayloadAttachments(record.Payload, attachments, includeAttachmentContent)
-
-	s.writeAuditBestEffort(ctx, AuditEvent{
-		VaultID:     row.VaultID,
-		WorkspaceID: row.WorkspaceID,
-		ActorType:   access.ActorType,
-		ActorID:     access.ActorID,
-		Action:      "read",
-		RecordID:    record.ID,
-		RecordType:  record.Type,
-		Outcome:     "allowed",
-	})
-
-	return record, nil
-}
-
-func (s *Store) CreateRecord(ctx context.Context, record *Record, access AccessContext) error {
-	if s.db == nil {
-		return ErrSecretStoreUnavailable
-	}
-	if record == nil {
-		return fmt.Errorf("vault: record is required")
-	}
-
-	access = access.normalized()
-	resolvedVaultID, err := s.resolveVaultID(ctx, record.VaultID)
-	if err != nil {
-		return err
-	}
-	record.VaultID = resolvedVaultID
-	record.Type = normalizeRecordType(record.Type)
-	record.WorkspaceID = strings.TrimSpace(record.WorkspaceID)
-	record.FolderPath, err = normalizeFolderPath(record.FolderPath)
-	if err != nil {
-		return err
-	}
-	record.Label = strings.TrimSpace(record.Label)
-	record.Source = strings.TrimSpace(record.Source)
-	record.RetentionPolicy = strings.TrimSpace(record.RetentionPolicy)
-	record.Tags = normalizeTags(record.Tags)
-	if record.ID == "" {
-		record.ID = uuid.New().String()
-	}
-	if record.Label == "" {
-		record.Label = "Untitled Vault Entry"
-	}
-	if len(record.Payload) == 0 {
-		record.Payload = json.RawMessage(`{}`)
-	}
-	cleanPayload, inlineAttachments, _, err := extractInlineRecordAttachments(record.Payload)
-	if err != nil {
-		return err
-	}
-	record.Payload = cleanPayload
-	if _, err := s.getVault(ctx, record.VaultID); err != nil {
-		return err
-	}
-	vaultDB, err := s.openVaultContentDB(ctx, record.VaultID)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = vaultDB.Close() }()
-
-	_, writeCapability := capabilitiesForRecordType(record.Type)
-	if err := s.authorizeAccess(ctx, access, record.VaultID, record.WorkspaceID, writeCapability, record.Type, "create", record.ID); err != nil {
-		return err
-	}
-
-	dek, err := s.ensureDataEncryptionKey(ctx, record.VaultID, true)
-	if err != nil {
-		return err
-	}
-
-	now := s.now()
-	record.CreatedAt = now
-	record.UpdatedAt = now
-
-	metadataNonce, metadataCiphertext, err := encryptRecordMetadata(dek, encryptedRecordMetadata{
-		Label:      record.Label,
-		FolderPath: record.FolderPath,
-		Tags:       record.Tags,
-	})
-	if err != nil {
-		return err
-	}
-	payloadNonce, payloadCiphertext, err := encryptBytes(dek, record.Payload)
-	if err != nil {
-		return fmt.Errorf("encrypt record payload: %w", err)
-	}
-
-	tx, err := vaultDB.BeginTx(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("begin vault record transaction: %w", err)
-	}
-	defer func() { _ = tx.Rollback() }()
-
-	if _, err := tx.ExecContext(ctx, `
-		INSERT INTO vault_records (
-			id, vault_id, type, workspace_id, source, retention_policy,
-			metadata_nonce, metadata_ciphertext, payload_nonce, payload_ciphertext,
-			created_at, updated_at
-		) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-	`, record.ID, record.VaultID, record.Type, record.WorkspaceID, record.Source, record.RetentionPolicy,
-		metadataNonce, metadataCiphertext, payloadNonce, payloadCiphertext, record.CreatedAt, record.UpdatedAt); err != nil {
-		return fmt.Errorf("insert vault record: %w", err)
-	}
-
-	if err := ensureFolderPathWithExecutor(ctx, tx, dek, record.VaultID, record.FolderPath, record.CreatedAt); err != nil {
-		return err
-	}
-
-	if len(inlineAttachments) > 0 {
-		for index := range inlineAttachments {
-			inlineAttachments[index].Attachment.CreatedAt = record.CreatedAt
-		}
-		if err := replaceRecordAttachmentsWithExecutor(ctx, tx, dek, record, inlineAttachments); err != nil {
-			return err
-		}
-	}
-
-	if err := tx.Commit(); err != nil {
-		return fmt.Errorf("commit vault record transaction: %w", err)
-	}
-
-	s.writeAuditBestEffort(ctx, AuditEvent{
-		VaultID:     record.VaultID,
-		WorkspaceID: record.WorkspaceID,
-		ActorType:   access.ActorType,
-		ActorID:     access.ActorID,
-		Action:      "create",
-		RecordID:    record.ID,
-		RecordType:  record.Type,
-		Outcome:     "allowed",
-	})
-
-	return nil
-}
-
-func (s *Store) UpdateRecord(ctx context.Context, id string, update RecordUpdate, access AccessContext) (*Record, error) {
-	if s.db == nil {
-		return nil, ErrSecretStoreUnavailable
-	}
-
-	row, err := s.getRecordRow(ctx, id)
-	if err != nil {
-		return nil, err
-	}
-	vaultDB, err := s.openVaultContentDB(ctx, row.VaultID)
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = vaultDB.Close() }()
-
-	access = access.normalized()
-	_, writeCapability := capabilitiesForRecordType(row.Type)
-	if err := s.authorizeAccess(ctx, access, row.VaultID, row.WorkspaceID, writeCapability, row.Type, "update", id); err != nil {
-		return nil, err
-	}
-
-	dek, err := s.ensureDataEncryptionKey(ctx, row.VaultID, false)
-	if err != nil {
-		return nil, err
-	}
-
-	record, err := decryptRecordRow(dek, row)
-	if err != nil {
-		return nil, err
-	}
-
-	if update.Type != nil {
-		record.Type = normalizeRecordType(*update.Type)
-	}
-	if update.WorkspaceID != nil {
-		record.WorkspaceID = strings.TrimSpace(*update.WorkspaceID)
-	}
-	if update.FolderPath != nil {
-		record.FolderPath, err = normalizeFolderPath(*update.FolderPath)
-		if err != nil {
-			return nil, err
-		}
-	}
-	if update.Label != nil {
-		record.Label = strings.TrimSpace(*update.Label)
-		if record.Label == "" {
-			record.Label = "Untitled Vault Entry"
-		}
-	}
-	if update.Tags != nil {
-		record.Tags = normalizeTags(*update.Tags)
-	}
-	if update.Source != nil {
-		record.Source = strings.TrimSpace(*update.Source)
-	}
-	if update.RetentionPolicy != nil {
-		record.RetentionPolicy = strings.TrimSpace(*update.RetentionPolicy)
-	}
-	if update.Payload != nil {
-		record.Payload = *update.Payload
-		if len(record.Payload) == 0 {
-			record.Payload = json.RawMessage(`{}`)
-		}
-	}
-	cleanPayload, inlineAttachments, attachmentsFieldPresent, err := extractInlineRecordAttachments(record.Payload)
-	if err != nil {
-		return nil, err
-	}
-	record.Payload = cleanPayload
-
-	_, targetWriteCapability := capabilitiesForRecordType(record.Type)
-	if record.Type != row.Type || record.WorkspaceID != row.WorkspaceID {
-		if err := s.authorizeAccess(ctx, access, row.VaultID, record.WorkspaceID, targetWriteCapability, record.Type, "update", id); err != nil {
-			return nil, err
-		}
-	}
-	record.UpdatedAt = s.now()
-
-	metadataNonce, metadataCiphertext, err := encryptRecordMetadata(dek, encryptedRecordMetadata{
-		Label:      record.Label,
-		FolderPath: record.FolderPath,
-		Tags:       record.Tags,
-	})
-	if err != nil {
-		return nil, err
-	}
-	payloadNonce, payloadCiphertext, err := encryptBytes(dek, record.Payload)
-	if err != nil {
-		return nil, fmt.Errorf("encrypt record payload: %w", err)
-	}
-
-	tx, err := vaultDB.BeginTx(ctx, nil)
-	if err != nil {
-		return nil, fmt.Errorf("begin vault record transaction: %w", err)
-	}
-	defer func() { _ = tx.Rollback() }()
-
-	result, err := tx.ExecContext(ctx, `
-		UPDATE vault_records
-		SET type = ?, workspace_id = ?, source = ?, retention_policy = ?, metadata_nonce = ?, metadata_ciphertext = ?,
-		    payload_nonce = ?, payload_ciphertext = ?, updated_at = ?
-		WHERE id = ?
-	`, record.Type, record.WorkspaceID, record.Source, record.RetentionPolicy, metadataNonce, metadataCiphertext,
-		payloadNonce, payloadCiphertext, record.UpdatedAt, id)
-	if err != nil {
-		return nil, fmt.Errorf("update vault record: %w", err)
-	}
-	if err := database.CheckRowsAffectedWithError(result, "vault_record", ErrRecordNotFound); err != nil {
-		return nil, err
-	}
-
-	if err := ensureFolderPathWithExecutor(ctx, tx, dek, record.VaultID, record.FolderPath, record.UpdatedAt); err != nil {
-		return nil, err
-	}
-
-	if attachmentsFieldPresent {
-		for index := range inlineAttachments {
-			inlineAttachments[index].Attachment.CreatedAt = record.UpdatedAt
-		}
-		if err := replaceRecordAttachmentsWithExecutor(ctx, tx, dek, record, inlineAttachments); err != nil {
-			return nil, err
-		}
-	}
-
-	if err := tx.Commit(); err != nil {
-		return nil, fmt.Errorf("commit vault record transaction: %w", err)
-	}
-
-	s.writeAuditBestEffort(ctx, AuditEvent{
-		VaultID:     record.VaultID,
-		WorkspaceID: record.WorkspaceID,
-		ActorType:   access.ActorType,
-		ActorID:     access.ActorID,
-		Action:      "update",
-		RecordID:    record.ID,
-		RecordType:  record.Type,
-		Outcome:     "allowed",
-	})
-
-	return record, nil
-}
-
-func (s *Store) DeleteRecord(ctx context.Context, id string, access AccessContext) error {
-	if s.db == nil {
-		return ErrSecretStoreUnavailable
-	}
-
-	row, err := s.getRecordRow(ctx, id)
-	if err != nil {
-		return err
-	}
-	vaultDB, err := s.openVaultContentDB(ctx, row.VaultID)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = vaultDB.Close() }()
-
-	access = access.normalized()
-	_, writeCapability := capabilitiesForRecordType(row.Type)
-	if err := s.authorizeAccess(ctx, access, row.VaultID, row.WorkspaceID, writeCapability, row.Type, "delete", id); err != nil {
-		return err
-	}
-
-	tx, err := vaultDB.BeginTx(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("begin vault record delete transaction: %w", err)
-	}
-	defer func() { _ = tx.Rollback() }()
-
-	if _, err := tx.ExecContext(ctx, `DELETE FROM vault_record_attachments WHERE record_id = ?`, id); err != nil {
-		return fmt.Errorf("delete vault record attachments: %w", err)
-	}
-
-	result, err := tx.ExecContext(ctx, `DELETE FROM vault_records WHERE id = ?`, id)
-	if err != nil {
-		return fmt.Errorf("delete vault record: %w", err)
-	}
-	if err := database.CheckRowsAffectedWithError(result, "vault_record", ErrRecordNotFound); err != nil {
-		return err
-	}
-
-	if err := tx.Commit(); err != nil {
-		return fmt.Errorf("commit vault record delete transaction: %w", err)
-	}
-
-	s.writeAuditBestEffort(ctx, AuditEvent{
-		VaultID:     row.VaultID,
-		WorkspaceID: row.WorkspaceID,
-		ActorType:   access.ActorType,
-		ActorID:     access.ActorID,
-		Action:      "delete",
-		RecordID:    row.ID,
-		RecordType:  row.Type,
-		Outcome:     "allowed",
-	})
-
-	return nil
-}
-
-func (s *Store) ListGrants(ctx context.Context, vaultID string, workspaceID string) ([]Grant, error) {
-	if s.db == nil {
-		return nil, ErrSecretStoreUnavailable
-	}
-
-	vaultID, err := s.resolveVaultID(ctx, vaultID)
-	if err != nil {
-		return nil, err
-	}
-	workspaceID = strings.TrimSpace(workspaceID)
-	if _, err := s.getVault(ctx, vaultID); err != nil {
-		return nil, err
-	}
-	vaultDB, err := s.openVaultContentDB(ctx, vaultID)
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = vaultDB.Close() }()
-	query := `
-		SELECT id, vault_id, workspace_id, actor_type, actor_id, capability, record_type, created_at, updated_at
-		FROM vault_grants
-		WHERE vault_id = ?
-		  AND (? = '' OR workspace_id = ?)
-		ORDER BY updated_at DESC
-	`
-	rows, err := vaultDB.QueryContext(ctx, query, vaultID, workspaceID, workspaceID)
-	if err != nil {
-		return nil, fmt.Errorf("query vault grants: %w", err)
-	}
-	defer func() { _ = rows.Close() }()
-
-	grants := make([]Grant, 0)
-	for rows.Next() {
-		var grant Grant
-		if err := rows.Scan(
-			&grant.ID,
-			&grant.VaultID,
-			&grant.WorkspaceID,
-			&grant.ActorType,
-			&grant.ActorID,
-			&grant.Capability,
-			&grant.RecordType,
-			&grant.CreatedAt,
-			&grant.UpdatedAt,
-		); err != nil {
-			return nil, fmt.Errorf("scan vault grant: %w", err)
-		}
-		grants = append(grants, grant)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("iterate vault grants: %w", err)
-	}
-	return grants, nil
-}
-
-func (s *Store) CreateGrant(ctx context.Context, grant *Grant) error {
-	if s.db == nil {
-		return ErrSecretStoreUnavailable
-	}
-	if grant == nil {
-		return fmt.Errorf("vault: grant is required")
-	}
-
-	grant.WorkspaceID = strings.TrimSpace(grant.WorkspaceID)
-	resolvedVaultID, err := s.resolveVaultID(ctx, grant.VaultID)
-	if err != nil {
-		return err
-	}
-	grant.VaultID = resolvedVaultID
-	grant.ActorType = normalizeActorType(grant.ActorType)
-	grant.ActorID = strings.TrimSpace(grant.ActorID)
-	grant.Capability = normalizeCapability(grant.Capability)
-	grant.RecordType = normalizeRecordType(grant.RecordType)
-	if grant.WorkspaceID == "" || grant.ActorID == "" || (grant.ActorType != ActorTypeAgent && grant.ActorType != ActorTypePlugin) || grant.Capability == "" {
-		return fmt.Errorf("vault: workspace_id, actor_type, actor_id, and capability are required")
-	}
-	if _, err := s.getVault(ctx, grant.VaultID); err != nil {
-		return err
-	}
-	vaultDB, err := s.openVaultContentDB(ctx, grant.VaultID)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = vaultDB.Close() }()
-
-	now := s.now()
-
-	var existingID string
-	var existingCreatedAt time.Time
-	err = vaultDB.QueryRowContext(ctx, `
-		SELECT id, created_at
-		FROM vault_grants
-		WHERE vault_id = ? AND workspace_id = ? AND actor_type = ? AND actor_id = ? AND capability = ? AND record_type = ?
-	`, grant.VaultID, grant.WorkspaceID, grant.ActorType, grant.ActorID, grant.Capability, grant.RecordType).Scan(&existingID, &existingCreatedAt)
-	switch {
-	case err == nil:
-		grant.ID = existingID
-		grant.CreatedAt = existingCreatedAt
-		grant.UpdatedAt = now
-		_, err = vaultDB.ExecContext(ctx, `
-			UPDATE vault_grants
-			SET updated_at = ?
-			WHERE id = ?
-		`, now, existingID)
-		if err != nil {
-			return fmt.Errorf("refresh vault grant: %w", err)
-		}
-	case errors.Is(err, sql.ErrNoRows):
-		if grant.ID == "" {
-			grant.ID = uuid.New().String()
-		}
-		grant.CreatedAt = now
-		grant.UpdatedAt = now
-		_, err = vaultDB.ExecContext(ctx, `
-			INSERT INTO vault_grants (
-				id, vault_id, workspace_id, actor_type, actor_id, capability, record_type, created_at, updated_at
-			) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)
-		`, grant.ID, grant.VaultID, grant.WorkspaceID, grant.ActorType, grant.ActorID, grant.Capability, grant.RecordType, grant.CreatedAt, grant.UpdatedAt)
-		if err != nil {
-			return fmt.Errorf("insert vault grant: %w", err)
-		}
-	default:
-		return fmt.Errorf("query existing vault grant: %w", err)
-	}
-
-	s.writeAuditBestEffort(ctx, AuditEvent{
-		VaultID:     grant.VaultID,
-		WorkspaceID: grant.WorkspaceID,
-		Action:      "grant.create",
-		RecordType:  grant.RecordType,
-		Outcome:     "allowed",
-		Details:     fmt.Sprintf(`{"actor_type":%q,"actor_id":%q,"capability":%q}`, grant.ActorType, grant.ActorID, grant.Capability),
-	})
-
-	return nil
-}
-
-func (s *Store) DeleteGrant(ctx context.Context, id string) error {
-	if s.db == nil {
-		return ErrSecretStoreUnavailable
-	}
-
-	id = strings.TrimSpace(id)
-	if id == "" {
-		return ErrGrantNotFound
-	}
-
-	vaults, err := s.listVaultCatalog(ctx)
-	if err != nil {
-		return err
-	}
-
-	var grant Grant
-	var vaultDB *sql.DB
-	for _, item := range vaults {
-		vaultDB, err = openExistingVaultFile(ctx, s.resolveVaultFileAbsolutePath(item.FilePath))
-		if err != nil {
-			if errors.Is(err, ErrVaultFileMissing) || errors.Is(err, ErrVaultFileCorrupt) {
-				continue
-			}
-			return fmt.Errorf("open vault content database: %w", err)
-		}
-
-		err = vaultDB.QueryRowContext(ctx, `
-			SELECT id, vault_id, workspace_id, actor_type, actor_id, capability, record_type, created_at, updated_at
-			FROM vault_grants
-			WHERE id = ?
-		`, id).Scan(
-			&grant.ID,
-			&grant.VaultID,
-			&grant.WorkspaceID,
-			&grant.ActorType,
-			&grant.ActorID,
-			&grant.Capability,
-			&grant.RecordType,
-			&grant.CreatedAt,
-			&grant.UpdatedAt,
-		)
-		if err == nil {
-			break
-		}
-		_ = vaultDB.Close()
-		vaultDB = nil
-		if errors.Is(err, sql.ErrNoRows) {
-			continue
-		}
-		return fmt.Errorf("get vault grant: %w", err)
-	}
-	if vaultDB == nil {
-		return ErrGrantNotFound
-	}
-	defer func() { _ = vaultDB.Close() }()
-
-	result, err := vaultDB.ExecContext(ctx, `DELETE FROM vault_grants WHERE id = ?`, grant.ID)
-	if err != nil {
-		return fmt.Errorf("delete vault grant: %w", err)
-	}
-	if err := database.CheckRowsAffectedWithError(result, "vault_grant", ErrGrantNotFound); err != nil {
-		return err
-	}
-
-	s.writeAuditBestEffort(ctx, AuditEvent{
-		VaultID:     grant.VaultID,
-		WorkspaceID: grant.WorkspaceID,
-		Action:      "grant.delete",
-		RecordType:  grant.RecordType,
-		Outcome:     "allowed",
-		Details:     fmt.Sprintf(`{"actor_type":%q,"actor_id":%q,"capability":%q}`, grant.ActorType, grant.ActorID, grant.Capability),
-	})
-
-	return nil
-}
-
-func (s *Store) Export(ctx context.Context, req ExportRequest) (*ExportBundle, error) {
-	if s.db == nil {
-		return nil, ErrSecretStoreUnavailable
-	}
-
-	req.WorkspaceID = strings.TrimSpace(req.WorkspaceID)
-	resolvedVaultID, err := s.resolveVaultID(ctx, req.VaultID)
-	if err != nil {
-		return nil, err
-	}
-	req.VaultID = resolvedVaultID
-	req.Password = strings.TrimSpace(req.Password)
-	if req.Password == "" {
-		return nil, ErrExportPasswordEmpty
-	}
-	selectedVault, err := s.getVault(ctx, req.VaultID)
-	if err != nil {
-		return nil, err
-	}
-
-	records, err := s.ListRecords(ctx, RecordFilter{VaultID: req.VaultID, WorkspaceID: req.WorkspaceID}, AccessContext{})
-	if err != nil {
-		return nil, err
-	}
-
-	fullRecords := make([]Record, 0, len(records))
-	for _, item := range records {
-		record, err := s.getRecord(ctx, item.ID, AccessContext{}, true)
-		if err != nil {
-			return nil, err
-		}
-		fullRecords = append(fullRecords, *record)
-	}
-
-	grants, err := s.ListGrants(ctx, req.VaultID, req.WorkspaceID)
-	if err != nil {
-		return nil, err
-	}
-
-	folders, err := s.ListFolders(ctx, req.VaultID)
-	if err != nil {
-		return nil, err
-	}
-
-	envelope := exportEnvelope{
-		VaultID:     req.VaultID,
-		VaultName:   selectedVault.Name,
-		ExportedAt:  s.now(),
-		WorkspaceID: req.WorkspaceID,
-		Folders:     folders,
-		Records:     fullRecords,
-		Grants:      grants,
-	}
-
-	salt, err := randomBytes(16)
-	if err != nil {
-		return nil, fmt.Errorf("generate export salt: %w", err)
-	}
-	derivedKey := derivePassphraseKey(req.Password, salt)
-	nonce, ciphertext, err := encryptJSON(derivedKey, envelope)
-	if err != nil {
-		return nil, fmt.Errorf("encrypt vault export: %w", err)
-	}
-
-	bundle := &ExportBundle{
-		Version:     1,
-		VaultID:     req.VaultID,
-		VaultName:   selectedVault.Name,
-		WorkspaceID: req.WorkspaceID,
-		Salt:        base64.StdEncoding.EncodeToString(salt),
-		Nonce:       nonce,
-		Ciphertext:  ciphertext,
-		ExportedAt:  envelope.ExportedAt,
-		RecordCount: len(fullRecords),
-		GrantCount:  len(grants),
-	}
-
-	s.writeAuditBestEffort(ctx, AuditEvent{
-		VaultID:     req.VaultID,
-		WorkspaceID: req.WorkspaceID,
-		Action:      "export",
-		Outcome:     "allowed",
-		Details:     fmt.Sprintf(`{"record_count":%d,"grant_count":%d}`, bundle.RecordCount, bundle.GrantCount),
-	})
-
-	return bundle, nil
-}
-
-func DecryptExportBundle(bundle ExportBundle, password string) ([]byte, error) {
-	password = strings.TrimSpace(password)
-	if password == "" {
-		return nil, ErrImportPasswordRequired
-	}
-	if strings.TrimSpace(bundle.Salt) == "" || strings.TrimSpace(bundle.Nonce) == "" || strings.TrimSpace(bundle.Ciphertext) == "" {
-		return nil, ErrImportBundleRequired
-	}
-
-	salt, err := base64.StdEncoding.DecodeString(bundle.Salt)
-	if err != nil {
-		return nil, ErrImportBundleInvalid
-	}
-	derivedKey := derivePassphraseKey(password, salt)
-	plaintext, err := decryptBytes(derivedKey, bundle.Nonce, bundle.Ciphertext)
-	if err != nil {
-		return nil, ErrImportPasswordInvalid
-	}
-	return plaintext, nil
-}
-
-func (s *Store) Import(ctx context.Context, req ImportRequest) (*ImportResult, error) {
-	if s.db == nil {
-		return nil, ErrSecretStoreUnavailable
-	}
-
-	req.Password = strings.TrimSpace(req.Password)
-	if req.Password == "" {
-		return nil, ErrImportPasswordRequired
-	}
-	if strings.TrimSpace(req.Bundle.Salt) == "" || strings.TrimSpace(req.Bundle.Nonce) == "" || strings.TrimSpace(req.Bundle.Ciphertext) == "" {
-		return nil, ErrImportBundleRequired
-	}
-
-	decrypted, err := DecryptExportBundle(req.Bundle, req.Password)
-	if err != nil {
-		return nil, err
-	}
-
-	var envelope exportEnvelope
-	if err := json.Unmarshal(decrypted, &envelope); err != nil {
-		return nil, ErrImportBundleInvalid
-	}
-
-	var targetVault Vault
-	createdVault := false
-
-	if strings.TrimSpace(req.TargetVaultID) != "" {
-		resolvedVaultID, err := s.resolveVaultID(ctx, req.TargetVaultID)
-		if err != nil {
-			return nil, err
-		}
-		targetVault, err = s.getVault(ctx, resolvedVaultID)
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		name := strings.TrimSpace(req.NewVaultName)
-		if name == "" {
-			name = strings.TrimSpace(envelope.VaultName)
-		}
-		if name == "" {
-			name = "Imported Vault"
-		}
-
-		item := Vault{
-			Name:        name,
-			Description: strings.TrimSpace(req.NewVaultDescription),
-		}
-		if err := s.CreateVault(ctx, &item, req.NewVaultPassword); err != nil {
-			return nil, err
-		}
-		targetVault = item
-		createdVault = true
-	}
-
-	importedRecords := 0
-	for _, item := range envelope.Folders {
-		folder := item
-		folder.ID = ""
-		folder.VaultID = targetVault.ID
-		if _, err := s.CreateFolder(ctx, &folder); err != nil {
-			return nil, err
-		}
-	}
-
-	for _, item := range envelope.Records {
-		record := item
-		record.ID = ""
-		record.VaultID = targetVault.ID
-		if strings.TrimSpace(record.Source) == "" {
-			record.Source = "import"
-		}
-		if err := s.CreateRecord(ctx, &record, AccessContext{}); err != nil {
-			return nil, err
-		}
-		importedRecords++
-	}
-
-	importedGrants := 0
-	if req.RestoreGrants {
-		for _, item := range envelope.Grants {
-			grant := item
-			grant.ID = ""
-			grant.VaultID = targetVault.ID
-			if err := s.CreateGrant(ctx, &grant); err != nil {
-				return nil, err
-			}
-			importedGrants++
-		}
-	}
-
-	targetVault, err = s.getVault(ctx, targetVault.ID)
-	if err != nil {
-		return nil, err
-	}
-
-	s.writeAuditBestEffort(ctx, AuditEvent{
-		VaultID: targetVault.ID,
-		Action:  "import",
-		Outcome: "allowed",
-		Details: fmt.Sprintf(`{"record_count":%d,"grant_count":%d,"created_vault":%t,"source_vault_name":%q}`, importedRecords, importedGrants, createdVault, envelope.VaultName),
-	})
-
-	return &ImportResult{
-		Vault:           targetVault,
-		CreatedVault:    createdVault,
-		SourceVaultName: strings.TrimSpace(envelope.VaultName),
-		RecordCount:     importedRecords,
-		GrantCount:      importedGrants,
-	}, nil
-}
-
-func (s *Store) ensureDataEncryptionKey(ctx context.Context, vaultID string, allowCreate bool) ([]byte, error) {
-	vaultID = normalizeVaultID(vaultID)
-	_ = allowCreate
-
-	s.mu.RLock()
-	if cached := s.cachedDEKs[vaultID]; len(cached) > 0 {
-		keyCopy := append([]byte(nil), cached...)
-		s.mu.RUnlock()
-		return keyCopy, nil
-	}
-	s.mu.RUnlock()
-
-	keyMaterial, err := s.getVaultKeyMaterial(ctx, vaultID)
-	if err != nil {
-		return nil, err
-	}
-	if keyMaterial.PasswordProtected {
-		return nil, ErrVaultLocked
-	}
-	if allowCreate {
-		return nil, ErrVaultKeyUnavailable
-	}
-	return nil, ErrVaultKeyUnavailable
-}
-
-func (s *Store) recordCount(ctx context.Context, vaultID string) (int, error) {
-	vaultID = normalizeVaultID(vaultID)
-	if vaultID == "" {
-		return 0, ErrVaultRequired
-	}
-	vaultDB, err := s.openVaultContentDB(ctx, vaultID)
-	if err != nil {
-		return 0, err
-	}
-	defer func() { _ = vaultDB.Close() }()
-
-	var count int
-	if err := vaultDB.QueryRowContext(ctx, `SELECT COUNT(*) FROM vault_records WHERE vault_id = ?`, vaultID).Scan(&count); err != nil {
-		return 0, fmt.Errorf("count vault records: %w", err)
-	}
-	return count, nil
-}
-
-func (s *Store) hasMatchingRecords(ctx context.Context, filter RecordFilter) (bool, error) {
-	filter.VaultID = normalizeVaultID(filter.VaultID)
-	filter.WorkspaceID = strings.TrimSpace(filter.WorkspaceID)
-	filter.Type = normalizeRecordType(filter.Type)
-	if filter.Type == "*" {
-		filter.Type = ""
-	}
-	vaultDB, err := s.openVaultContentDB(ctx, filter.VaultID)
-	if err != nil {
-		return false, err
-	}
-	defer func() { _ = vaultDB.Close() }()
-
-	var exists int
-	err = vaultDB.QueryRowContext(ctx, `
-		SELECT 1
-		FROM vault_records
-		WHERE vault_id = ?
-		  AND (? = '' OR workspace_id = ?)
-		  AND (? = '' OR type = ?)
-		LIMIT 1
-	`, filter.VaultID, filter.WorkspaceID, filter.WorkspaceID, filter.Type, filter.Type).Scan(&exists)
-	switch {
-	case errors.Is(err, sql.ErrNoRows):
-		return false, nil
-	case err != nil:
-		return false, fmt.Errorf("query matching vault records: %w", err)
-	default:
-		return exists == 1, nil
-	}
-}
-
 func (s *Store) getVault(ctx context.Context, vaultID string) (Vault, error) {
 	item, err := s.getVaultCatalog(ctx, vaultID)
 	if err != nil {
@@ -1926,197 +920,6 @@ func (s *Store) getVault(ctx context.Context, vaultID string) (Vault, error) {
 	}
 	item.RecordCount = recordCount
 	return item, nil
-}
-
-func (s *Store) getRecordRow(ctx context.Context, id string) (recordRow, error) {
-	id = strings.TrimSpace(id)
-	if id == "" {
-		return recordRow{}, ErrRecordNotFound
-	}
-
-	vaults, err := s.listVaultCatalog(ctx)
-	if err != nil {
-		return recordRow{}, err
-	}
-
-	for _, item := range vaults {
-		vaultDB, err := openExistingVaultFile(ctx, s.resolveVaultFileAbsolutePath(item.FilePath))
-		if err != nil {
-			if errors.Is(err, ErrVaultFileMissing) || errors.Is(err, ErrVaultFileCorrupt) {
-				continue
-			}
-			return recordRow{}, fmt.Errorf("open vault content database: %w", err)
-		}
-
-		row, err := getRecordRowWithExecutor(ctx, vaultDB, id)
-		_ = vaultDB.Close()
-		if err == nil {
-			return row, nil
-		}
-		if errors.Is(err, ErrRecordNotFound) {
-			continue
-		}
-		return recordRow{}, err
-	}
-
-	return recordRow{}, ErrRecordNotFound
-}
-
-func getRecordRowWithExecutor(ctx context.Context, executor attachmentSQLExecutor, id string) (recordRow, error) {
-	var row recordRow
-	err := executor.QueryRowContext(ctx, `
-		SELECT id, vault_id, type, workspace_id, source, retention_policy,
-		       metadata_nonce, metadata_ciphertext, payload_nonce, payload_ciphertext,
-		       created_at, updated_at
-		FROM vault_records
-		WHERE id = ?
-	`, id).Scan(
-		&row.ID,
-		&row.VaultID,
-		&row.Type,
-		&row.WorkspaceID,
-		&row.Source,
-		&row.RetentionPolicy,
-		&row.MetadataNonce,
-		&row.MetadataCiphertext,
-		&row.PayloadNonce,
-		&row.PayloadCiphertext,
-		&row.CreatedAt,
-		&row.UpdatedAt,
-	)
-	if errors.Is(err, sql.ErrNoRows) {
-		return recordRow{}, ErrRecordNotFound
-	}
-	if err != nil {
-		return recordRow{}, fmt.Errorf("get vault record: %w", err)
-	}
-	return row, nil
-}
-
-func (s *Store) authorizeList(ctx context.Context, access AccessContext, filter RecordFilter) error {
-	if !access.requiresGrant() {
-		return nil
-	}
-	if !isValidAccessActor(access) {
-		s.writeAuditBestEffort(ctx, AuditEvent{
-			VaultID:     filter.VaultID,
-			WorkspaceID: access.WorkspaceID,
-			ActorType:   access.ActorType,
-			ActorID:     access.ActorID,
-			Action:      "list",
-			Outcome:     "denied",
-			Details:     `{"reason":"invalid_actor"}`,
-		})
-		return ErrPermissionDenied
-	}
-	if access.WorkspaceID == "" || (filter.WorkspaceID != "" && filter.WorkspaceID != access.WorkspaceID) {
-		s.writeAuditBestEffort(ctx, AuditEvent{
-			VaultID:     filter.VaultID,
-			WorkspaceID: access.WorkspaceID,
-			ActorType:   access.ActorType,
-			ActorID:     access.ActorID,
-			Action:      "list",
-			Outcome:     "denied",
-			Details:     `{"reason":"workspace_scope_missing"}`,
-		})
-		return ErrPermissionDenied
-	}
-	if !s.hasAnyGrant(ctx, filter.VaultID, access) {
-		s.writeAuditBestEffort(ctx, AuditEvent{
-			VaultID:     filter.VaultID,
-			WorkspaceID: access.WorkspaceID,
-			ActorType:   access.ActorType,
-			ActorID:     access.ActorID,
-			Action:      "list",
-			Outcome:     "denied",
-			Details:     `{"reason":"persistent_grant_missing"}`,
-		})
-		return ErrPermissionDenied
-	}
-	return nil
-}
-
-func (s *Store) authorizeAccess(ctx context.Context, access AccessContext, vaultID string, workspaceID string, capability Capability, recordType string, action string, recordID string) error {
-	if !access.requiresGrant() {
-		return nil
-	}
-	if !isValidAccessActor(access) || access.WorkspaceID == "" || workspaceID == "" || access.WorkspaceID != workspaceID {
-		s.writeAuditBestEffort(ctx, AuditEvent{
-			VaultID:     vaultID,
-			WorkspaceID: workspaceID,
-			ActorType:   access.ActorType,
-			ActorID:     access.ActorID,
-			Action:      action,
-			RecordID:    recordID,
-			RecordType:  recordType,
-			Outcome:     "denied",
-			Details:     `{"reason":"workspace_scope_missing"}`,
-		})
-		return ErrPermissionDenied
-	}
-	if !s.hasGrant(ctx, vaultID, access, workspaceID, capability, recordType) {
-		s.writeAuditBestEffort(ctx, AuditEvent{
-			VaultID:     vaultID,
-			WorkspaceID: workspaceID,
-			ActorType:   access.ActorType,
-			ActorID:     access.ActorID,
-			Action:      action,
-			RecordID:    recordID,
-			RecordType:  recordType,
-			Outcome:     "denied",
-			Details:     fmt.Sprintf(`{"reason":"persistent_grant_missing","capability":%q}`, capability),
-		})
-		return ErrPermissionDenied
-	}
-	return nil
-}
-
-func (s *Store) hasAnyGrant(ctx context.Context, vaultID string, access AccessContext) bool {
-	vaultDB, err := s.openVaultContentDB(ctx, vaultID)
-	if err != nil {
-		return false
-	}
-	defer func() { _ = vaultDB.Close() }()
-	return hasAnyGrantWithExecutor(ctx, vaultDB, vaultID, access)
-}
-
-func hasAnyGrantWithExecutor(ctx context.Context, executor attachmentSQLExecutor, vaultID string, access AccessContext) bool {
-	var exists int
-	err := executor.QueryRowContext(ctx, `
-		SELECT 1
-		FROM vault_grants
-		WHERE vault_id = ? AND workspace_id = ? AND actor_type = ? AND actor_id = ?
-		LIMIT 1
-	`, normalizeVaultID(vaultID), access.WorkspaceID, access.ActorType, access.ActorID).Scan(&exists)
-	return err == nil && exists == 1
-}
-
-func (s *Store) hasGrant(ctx context.Context, vaultID string, access AccessContext, workspaceID string, capability Capability, recordType string) bool {
-	vaultDB, err := s.openVaultContentDB(ctx, vaultID)
-	if err != nil {
-		return false
-	}
-	defer func() { _ = vaultDB.Close() }()
-	return hasGrantWithExecutor(ctx, vaultDB, vaultID, access, workspaceID, capability, recordType)
-}
-
-func hasGrantWithExecutor(ctx context.Context, executor attachmentSQLExecutor, vaultID string, access AccessContext, workspaceID string, capability Capability, recordType string) bool {
-	recordType = normalizeRecordType(recordType)
-	capability = normalizeCapability(capability)
-
-	var exists int
-	err := executor.QueryRowContext(ctx, `
-		SELECT 1
-		FROM vault_grants
-		WHERE vault_id = ?
-		  AND workspace_id = ?
-		  AND actor_type = ?
-		  AND actor_id = ?
-		  AND capability = ?
-		  AND (record_type = '*' OR record_type = ?)
-		LIMIT 1
-	`, normalizeVaultID(vaultID), workspaceID, access.ActorType, access.ActorID, capability, recordType).Scan(&exists)
-	return err == nil && exists == 1
 }
 
 func (s *Store) writeAuditBestEffort(ctx context.Context, event AuditEvent) {
@@ -2160,182 +963,6 @@ func (s *Store) writeAuditBestEffort(ctx context.Context, event AuditEvent) {
 		event.RecordID, event.RecordType, event.Outcome, event.Details, event.CreatedAt); err != nil {
 		logger.Warn("Failed to write vault audit event", logger.Fields{"error": err})
 	}
-}
-
-func scanRecordRow(rows interface {
-	Scan(dest ...any) error
-}) (recordRow, error) {
-	var row recordRow
-	if err := rows.Scan(
-		&row.ID,
-		&row.VaultID,
-		&row.Type,
-		&row.WorkspaceID,
-		&row.Source,
-		&row.RetentionPolicy,
-		&row.MetadataNonce,
-		&row.MetadataCiphertext,
-		&row.PayloadNonce,
-		&row.PayloadCiphertext,
-		&row.CreatedAt,
-		&row.UpdatedAt,
-	); err != nil {
-		return recordRow{}, fmt.Errorf("scan vault record: %w", err)
-	}
-	return row, nil
-}
-
-func decryptRecordRow(dek []byte, row recordRow) (*Record, error) {
-	metadata, err := decryptRecordMetadata(dek, row.MetadataNonce, row.MetadataCiphertext)
-	if err != nil {
-		return nil, err
-	}
-	payload, err := decryptBytes(dek, row.PayloadNonce, row.PayloadCiphertext)
-	if err != nil {
-		return nil, ErrMalformedRecord
-	}
-
-	return &Record{
-		ID:              row.ID,
-		VaultID:         row.VaultID,
-		Type:            row.Type,
-		WorkspaceID:     row.WorkspaceID,
-		FolderPath:      metadata.FolderPath,
-		Label:           metadata.Label,
-		Tags:            metadata.Tags,
-		Source:          row.Source,
-		RetentionPolicy: row.RetentionPolicy,
-		Payload:         json.RawMessage(payload),
-		CreatedAt:       row.CreatedAt,
-		UpdatedAt:       row.UpdatedAt,
-	}, nil
-}
-
-func encryptRecordMetadata(dek []byte, metadata encryptedRecordMetadata) (string, string, error) {
-	normalizedFolderPath, err := normalizeFolderPath(metadata.FolderPath)
-	if err != nil {
-		return "", "", err
-	}
-	metadata.FolderPath = normalizedFolderPath
-	metadata.Tags = normalizeTags(metadata.Tags)
-	return encryptJSON(dek, metadata)
-}
-
-func decryptRecordMetadata(dek []byte, nonceB64 string, ciphertextB64 string) (encryptedRecordMetadata, error) {
-	var metadata encryptedRecordMetadata
-	if err := decryptJSON(dek, nonceB64, ciphertextB64, &metadata); err != nil {
-		return encryptedRecordMetadata{}, ErrMalformedRecord
-	}
-	normalizedFolderPath, err := normalizeFolderPath(metadata.FolderPath)
-	if err != nil {
-		return encryptedRecordMetadata{}, ErrMalformedRecord
-	}
-	metadata.FolderPath = normalizedFolderPath
-	metadata.Tags = normalizeTags(metadata.Tags)
-	return metadata, nil
-}
-
-func isValidAccessActor(access AccessContext) bool {
-	access = access.normalized()
-	if access.ActorID == "" {
-		return false
-	}
-	return access.ActorType == ActorTypeAgent || access.ActorType == ActorTypePlugin
-}
-
-func (s *Store) hasCachedDEK(vaultID string) bool {
-	vaultID = normalizeVaultID(vaultID)
-	if vaultID == "" {
-		return false
-	}
-
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return len(s.cachedDEKs[vaultID]) > 0
-}
-
-func (s *Store) getVaultKeyMaterial(ctx context.Context, vaultID string) (vaultKeyMaterial, error) {
-	selectedVault, err := s.getVault(ctx, vaultID)
-	if err != nil {
-		return vaultKeyMaterial{}, err
-	}
-	if strings.TrimSpace(selectedVault.FilePath) == "" {
-		return vaultKeyMaterial{}, ErrVaultKeyUnavailable
-	}
-
-	vaultFileDB, err := openExistingVaultFile(ctx, s.resolveVaultFileAbsolutePath(selectedVault.FilePath))
-	if err != nil {
-		return vaultKeyMaterial{}, fmt.Errorf("open vault file: %w", err)
-	}
-	defer func() { _ = vaultFileDB.Close() }()
-
-	metadata, err := loadVaultFileMetadata(ctx, vaultFileDB, selectedVault.ID)
-	if err != nil {
-		return vaultKeyMaterial{}, err
-	}
-
-	material := vaultKeyMaterial{
-		Salt:              metadata.KeySalt,
-		Nonce:             metadata.KeyNonce,
-		Ciphertext:        metadata.KeyCiphertext,
-		PasswordProtected: strings.TrimSpace(metadata.KeyCiphertext) != "",
-	}
-	if !material.PasswordProtected {
-		return vaultKeyMaterial{}, ErrVaultKeyUnavailable
-	}
-	return material, nil
-}
-
-func passwordProtectedStoreStatus(locked bool) StoreStatus {
-	return StoreStatus{
-		Backend:   BackendVaultPassword,
-		Available: true,
-		Writable:  true,
-		Locked:    locked,
-		Message:   "using this vault's password-protected encryption key",
-	}
-}
-
-func validateVaultPassword(password string) error {
-	if strings.TrimSpace(password) == "" {
-		return ErrVaultPasswordRequired
-	}
-	return nil
-}
-
-func wrapVaultDataEncryptionKey(password string, dek []byte) (string, string, string, error) {
-	salt, err := randomBytes(16)
-	if err != nil {
-		return "", "", "", fmt.Errorf("generate vault password salt: %w", err)
-	}
-
-	derivedKey := derivePassphraseKey(password, salt)
-	nonce, ciphertext, err := encryptBytes(derivedKey, dek)
-	if err != nil {
-		return "", "", "", fmt.Errorf("encrypt vault data encryption key: %w", err)
-	}
-
-	return base64.StdEncoding.EncodeToString(salt), nonce, ciphertext, nil
-}
-
-func unwrapVaultDataEncryptionKey(password string, material vaultKeyMaterial) ([]byte, error) {
-	if !material.PasswordProtected {
-		return nil, ErrVaultKeyUnavailable
-	}
-
-	salt, err := base64.StdEncoding.DecodeString(material.Salt)
-	if err != nil {
-		return nil, ErrVaultKeyUnavailable
-	}
-	derivedKey := derivePassphraseKey(password, salt)
-	dek, err := decryptBytes(derivedKey, material.Nonce, material.Ciphertext)
-	if err != nil {
-		return nil, ErrVaultPasswordInvalid
-	}
-	if len(dek) != int(argonKeyLength) {
-		return nil, ErrVaultKeyUnavailable
-	}
-	return dek, nil
 }
 
 func (s *Store) resolveVaultID(ctx context.Context, vaultID string) (string, error) {
